+++ packages/angular_devkit/build_angular/src/builders/karma/tests/options/styles_spec.ts
@@ -7,10 +7,14 @@
  */
 
 import { execute } from '../../index';
-import { BASE_OPTIONS, KARMA_BUILDER_INFO, describeBuilder } from '../setup';
+import { BASE_OPTIONS, KARMA_BUILDER_INFO, describeKarmaBuilder } from '../setup';
 
-describeBuilder(execute, KARMA_BUILDER_INFO, (harness) => {
+describeKarmaBuilder(execute, KARMA_BUILDER_INFO, (harness, setupTarget) => {
   describe('Option: "styles"', () => {
+    beforeEach(async () => {
+      await setupTarget(harness);
+    });
+
     it(`processes 'styles.css' styles`, async () => {
       await harness.writeFiles({
         'src/styles.css': 'p {display: none}',
@@ -130,7 +134,9 @@ describeBuilder(execute, KARMA_BUILDER_INFO, (harness) => {
       expect(logs).toContain(
         jasmine.objectContaining({
           level: 'error',
-          message: jasmine.stringMatching(`Can't resolve 'src/test-style-a.css'`),
+          message: jasmine.stringMatching(
+            /(Can't|Could not) resolve ['"]src\/test-style-a.css['"]/,
+          ),
         }),
       );
     });